refactor(workspace): boucle 13 étape 1 — extract entities to crate simple-smtp-domain
Établit le Cargo workspace en 2 crates:
- racine simple-smtp-server (application + infrastructure + binaires)
- crates/domain simple-smtp-domain (entités métier, ex-src/entities.rs)

Approche non-disruptive: src/entities.rs devient un simple 'pub use
simple_smtp_domain::*;' qui préserve tous les call-sites (`use crate::entities::X`)
sans réécriture des 48 fichiers.

Étapes suivantes possibles (hors PR):
- Extraire logic/ vers crates/application (couche use-cases)
- Extraire mongo_adapter + smtp_client vers crates/infrastructure

Métriques spike: 9 use crate:: dans le code, structure déjà découpée,
extraction incrémentale sûre.
