+++ crates/domain/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "simple-smtp-domain"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+mongodb = "3.8.0"
+serde = { version = "1.0.229", features = ["derive"] }
+chrono = { version = "0.4.38", features = ["serde"] }
